Correctly rename kebab-case the files
