+++ src/app/app/virtual-lab/(free)/explore/(interactive)/interactive/(data)/model/circuit/[id]/page.tsx
@@ -1,4 +1,4 @@
-import CircuitDetails from '@/components/explore-section/Circuit/DetailView/MainDetailViewCore';
+import CircuitDetails from '@/components/explore-section/Circuit/DetailView/main-detail-view-core';
 
 export default function CircuitDetailPage() {
   return <CircuitDetails />;
