+++ src/app/app/virtual-lab/(free)/explore/(interactive)/interactive/(data)/model/circuit/page.tsx
@@ -1,4 +1,4 @@
-import CircuitsListingPageComponent from '@/components/explore-section/Circuit/global/CircuitsListingPageComponent';
+import CircuitsListingPageComponent from '@/components/explore-section/Circuit/global/circuits-listing-page-component';
 
 export default function ExploreModelCircuitListingPage() {
   return <CircuitsListingPageComponent />;
