+++ src/components/explore-section/Circuit/DetailView/header-detail-view.tsx
