+++ src/components/explore-section/Circuit/DetailView/sections/section-content-bloc.tsx
