+++ src/components/explore-section/Circuit/DetailView/sections/section-tabs.tsx
