+++ src/components/explore-section/Circuit/global/search-bar.tsx
